Add files via upload
Here is where I did the Calculations for the Regular Polygons. And here is working code for a function that takes an initializer list of boxes and finds the maximum dimensions from that list. 

Not to be used for the final project, but May help to understand what I am trying to do.